Changes to the Rendering API

+++ SpellcoreRenderer/source/core/components/SpellcoreRenderDataTypes.h
@@ -6,7 +6,7 @@
 namespace AnalyticalApproach::Spellcore
 {
     using SCRenderTargetHandle = std::uint32_t;
-    using SCRenderPassHandle = std::uint8_t;
+    using SCRenderPassHandle = std::uint64_t;
     using SCMaterialHandle = std::uint32_t; 
     using SCTextureHandle = std::uint32_t;
     using SCShaderHandle = std::uint16_t;
